REVERTME [VPG]: drm/i915: Add IS_BDW_GT3 macro.
It will be usefull to specify w/a that affects only BDW GT3.

v2: It is originally FROM_UPSTREAM but it uses INTEL_DEVID(dev) macro
which is not yet merged in our tree.
To fix compile errors and to continue with the rebase of other workaround
patches I have modified it to suit current convention hence it is a REVERTME.
Once we pull-in the upstream version this can be reverted.

For: VIZ-4092
Change-Id: I72b2c26500844a22e4ed1e5a3b4bea65b77f3486
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Upstream-repo: http://cgit.freedesktop.org/drm-intel/log/?h=drm-intel-next-queued
Upstream-commit: a0fcbd95354e5ac2400be74077cc668c91429ea7
Reviewed-by: Jesse Barnes <jbarnes@virtuousgeek.org>
Signed-off-by: Arun Siluvery <arun.siluvery@linux.intel.com>