REVERTME [VPG]: drm/i915: Program DDL, Water marks and DSPARB as per atomicity
DDL register is a single buffer register, hence any write to the
register will get into affect immediately. So during calculation if
there is a change in the calculated ddl value then write a value of '0'
to the DDL register and program the proper calculated value on the next
vblank, i.e in the flip done interrupt handler. Display arbitration control
register is programmed as per atomicity. REVERTME as some more changes need
to be done.

Issue: GMIN-2035
Change-Id: I459c46602297346418cb85c35779d3ec3a3e3071
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>
Signed-off-by: Vijay Purushothaman <vijay.a.purushothaman@intel.com>
Signed-off-by: Vidya Srinivas <vidya.srinivas@intel.com>
Signed-off-by: Sangani, Suryanarayana R <suryanarayana.r.sangani@intel.com>