REVERTME [VPG]: drm/i915: Skip wait for Gfx, if Gfx clock is already On
If GFX clk is already ON during suspend. Skip waiting for GFX to go down
and then forcing GFX clk on again. Instead return success to continue
with runtime suspend operation.

We skip the GFX  check for cherryview only.

This patch is REVERTME, With latest BIOS, we see GFX CLK On,
when we enter runtime suspend. Revert the patch once fixed.

Issue: GMINL-1917
Change-Id: I8168c2c719cb43e727357b945cfbfbdc65631f58
Signed-off-by: Deepak S <deepak.s@intel.com>