FROM_UPSTREAM [VPG]: drm/i915: Only mark as map-and-fenceable when bound into the GGTT
We use the obj->map_and_fenceable hint for when we already have a
valid mapping of this object in the aperture. This hint can only apply
to the GGTT and not to the aliasing-ppGTT. One user of the hint is
execbuffer relocation, which began to fail when it tried to follow the
hint and perform the relocate through the non-existent GGTT mapping.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=85671
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>

Backport note, replaced vma->bound to obj->has_global_gtt_mapping as
this is a fairly new change that is not present in GMINL.

Change-Id: I38659572f26f35693df01d5cd59b90f4bce01254
Upstream-repo: git://anongit.freedesktop.org/drm-intel
Upstream-commit: ef79e17ccef32599f3cec3842f81d70117181291
Upstream-as-of-tag: na
For: GMINL-4004
Signed-off-by: Michel Thierry <michel.thierry@intel.com>