USB IPV6 Mac address causes wrong router IPV6 configuration
When USB CDC-NCM driver is initialized, a MAC address is defines for this Ethernet interface.
In IPV6 stack, any Ethernet interface creates automatically a link/local address based on MAC.
Even if UE receives a new link/local from Network, IPV6 stack generates or uses automatically
the MAC address link/local insetad of the link/local coming from Network.
This could leads to wrong IPV6 Address assignement and failure in some test cases,
And it could confuse the network because several UE will do a router sollicitation
using same link/local (because the MAC address doesn't change for this cdc interface.

The fix consist to either reset the auto_linklocal (but not supported in theis kernel) or
remove or undefined the MAC address. To do so, we define the MAC length address to 1 like
many other network device driver does.

Change-Id: I52ee3deda05d6673217b1b49d8b81987e5837f56
Signed-off-by: feng wang <feng.a.wang@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Tested-by: Wang, Ran <ran.wang@intel.com>
Signed-off-by: Li Aixiong <aixiong.li@intel.com>