sched: Precisely accumulate periods in update_entity_runnable_avg
[ This patch has already been merged by the Linux kernel master,
see commit 9d89c257dfb9c51a532d693]

The current amount of time already accumulated against next period is
determined by:

  delta_w = sa->runnable_avg_period % 1024

Considering the runnable_avg_period is the sum of an infinite series, this
method looks gross. Precisely ccounting the period is not hard, just use a
variable period_contrib to record it.

The current runnable argument is either 1 or 0, indicating whether this entity
is runnable or not. In order to count the number of runnables as well, change
how runnable_avg_sum is accumulated.

Change-Id: Iec0ab7567b4ccffd9ffb66adcb637ab535004310
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-9774
Signed-off-by: Yuyang Du <yuyang.du@intel.com>
Signed-off-by: Srinidhi Kasagar <srinidhi.kasagar@intel.com>