FOR_UPSTREAM [VPG]: drm/i915: fix ULPS sequence for MIPI
during MIPI enabling, AFE latch should be enabled only after
2.5 ms later entering ULPS state, otherwise incorrect LP
state can be observed by the panel. Also during ULPS exit this
condition has to be taken care.

In the current code, device ready bit is cleared as part of ULPS
exit sequence which is wrong. fixing this issue in this patch.

In dual link, ULPS entry exit enter sequence is done ony by one for
each pipe. changing the logic to do ULPS enter for both the pipes
and then followed by exit and enter for both the pipes so that when
AFE latch is disabled, there is more chance to set the LP state as
LP-11.

Issue: GMINL-1644
Change-Id: I1374547b1519cd0a162a090f04e5d1179a70b157
Signed-off-by: Yogesh Mohan Marimuthu <yogesh.mohan.marimuthu@intel.com>