REVERTME [VPG]: drm/i915: send one frame after enabling mipi cmd mode
If MIPI is operated in command mode, and after display reset. if
not even one frame is sent after enabling the pipe and then if it
is disabled, pipe is getting stuck. This patch will fix this issue
by sending one frame after enabling the pipe.

REVERTME:
There should not be a case where there is mode set and no frames
are sent. Need to find out why the user space was doing this and
avoid it.

Issue: GMINL-1644
Change-Id: I5e3a59b47216aebdefc900f56f06388a06b54599
Signed-off-by: Yogesh Mohan Marimuthu <yogesh.mohan.marimuthu@intel.com>