usb/dwc3-device: fix super speed enumeration issue for CHT B1/C0
There's a silicon issue on CHT that everytime DWC3 enters and then
exits D0i3, bit 11 of GUSB3PIPECTL reg is cleared to 0, which leads
to device enumerated as a high speed one by PC's USB3 port.

This patch utilizes pci quirk method to set the bit when DWC3 is
resumed back as a workaroud.

Change-Id: Ice5f4ef5b870ffbc960134e1824ba4f6662a5d0d
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-9513
Signed-off-by: Jiebing Li <jiebing.li@intel.com>