MUST_REBASE [VPG]: drm/i915: Allow DPST register access when system in suspend
When system was in suspend and when we tried to read DPST status, we end
up reading wrong value & resulting in showing invalid data which is
causing GFT test it fail.

Since DPST registers is part Display unit, forcewake the Display unit to
read registers in proper state. Only Display 2D will be power up not the
panel, after the register read system goes back to suspend.

MUST_REBASE: DPST will not be upstreamed.

Issue: GMIN-3624
Change-Id: Ie5d62c15f9b5f3f78b8a43f0b58ac2216278e19d
Signed-off-by: Deepak S <deepak.s@intel.com>