MUST_REBASE [VPG]: adf/intel/vlv: Introduce vlv_pipeline
vlv_pipeline is a superset of vlv_pipe, vlv_port, vlv_pll, vlv_pm  and planes.
This patch adds changes required to plug in this new design across the
existing code and adds new files for pipeline as well.

MUST_REBASE: Since Google ADF framework is not in upstream and also
we cannot have two display drivers, other one being i915, we need to
work on i915 and ADF convergence path before it can be upstreamed.

For: GMINL-5213
Change-Id: I550c9c9f3380cd24561ea1a2d6995c590ee93207
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Signed-off-by: Deepak M <m.deepak@intel.com>