SQUASHME! [VPG]: drm/i915: Fix for stalling on cancelled fences
The scheduler supports batch buffers being dependent upon native sync
points. If a sync point is not yet signalled then that batch buffer
will not be submitted. If all remaining buffers are blocked by sync
points then the scheduler will sleep on those sync points.

Unfortunately, the test for whether a given batch buffer's sync point
was signaled or not did not take errored sync points into account.
This patch corrects the test so that it no longer waits forever if a
sync point fails for any reason.

This patch could be squashed in to the one that originally added the
scheduler:
  commit c64b5c1f236ee9c9362f06a675b4bc906ad3cb04
  Author: John Harrison <John.C.Harrison@Intel.com>
  Date:   Thu Apr 10 10:58:56 2014 +0100
  Subject: FOR_UPSTREAM [VPG]: drm/i915: Added trace points to scheduler
  Change-Id: I9886390cfc7897bc1faf50a104bc651d8baed8a5

Change-Id: Ib9d3cb86c3f9d184645b5cd57f76e8bb3266ae78
For: VIZ-1587
For: VIZ-4741
Signed-off-by: John Harrison <John.C.Harrison@Intel.com>