SQUASHME! [VPG]: drm/i915: Reduce context dependency severity in scheduler
The scheduler keeps batches within the same context in order. However,
it was doing this based on the application level context (which can
span multiple rings) not the hardware level context (which is per
ring). That is, it was forcing in order execution even when batch
buffers were submitted across different rings. This is a change of
behaviour. Before the scheduler, only intra-ring ordering was
enforced. Some of the validation tests started failing or complaining
of unexpected stalls due to this change of behvaiour.

This patch adds in a ring test to the context dependency check, thus
restoring the original behaviour.

This patch could be squashed in to the original scheduler patch that
added the dependency checking:
  commit 130244eba1ea2e5744455be5b6f06afdeece095e
  Author: John Harrison <John.C.Harrison@Intel.com>
  Date:   Tue Apr 1 16:27:39 2014 +0100
  Subject: FOR_UPSTREAM [VPG]: drm/i915: Start of GPU scheduler
  Change-Id: I1e08f59e650a3c2bbaaa9de7627da33849b06106

Change-Id: I42e94d2ae1c22cfa5a030103dd1863d6790a1ec2
For: VIZ-1587
For: VIZ-4741
Signed-off-by: John Harrison <John.C.Harrison@Intel.com>