FOR_UPSTREAM [VPG]: drm/i915: Generic DRRS state Machine
DRRS is a power saving feature, which will refresh the display
at the lowest supported refresh rate, based on the rate of change
of display content to be rendered.

This patch implements the Generic state machine for the Idleness DRRS.
Idleness DRRS is nothing but, when the content of the Display is not
changed for a certain duration, refresh rate will be set to the
minimum vrefresh supported by panel.

This will be extended for media playback DRRS. i.e based on the
display content's Frames Per Second, required lowest vrefresh to
support the FPS will be used to render the data. This involves
the Userspace and Kernel interaction.

DRRS Support for the encoders like DSI and eDP can be implemented
as a separate module and get registered with this generic DRRS stack.
Hence extending to new encoder and platform is made easy.

Issue: GMINL-8115
Change-Id: I89c7250ff6cb63579b9dcd05a2cd1fd9d8d43ad3
Signed-off-by: Ramalingam C <ramalingam.c@intel.com>