i2c: designware: check result of i2c_dw_enable/i2c_dw_xfer_init
We need to deal with the case that en/disable adapter failed or
i2c_dw_xfer_init failed. So, we change the prototype of the
function and re-initialize the i2c controller if the fail case
happens.
We also limit the time to be 10 to get notice earlier. The
average time that we've recorded is 2.

Change-Id: I007fbb00da12b2808f763a60612cbc10cf68fc57
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-30930
Tracked-On: https://jira01.devtools.intel.com/browse/AKP-1824
Signed-off-by: Jun Zhang <jun.zhang@intel.com>
Signed-off-by: Chuansheng Liu <chuansheng.liu@intel.com>
Signed-off-by: Fei Li <fei.li@intel.com>
Signed-off-by: Xiang Wang <xiang.a.wang@intel.com>