tty/n_gsm.c: clear gsm_mux[] entry
If the gsmtty is still used by some process, we could
not just clear gsm_mux[gsm->num]. Clear it when mux is
being free. Otherwise we may reuse an old-but-still-opend
device and hit a crash when userspace close the gsmtty.
This patch alos fix the race between gsmld_config and gsmld_open
As gsmld_config may call gsm_cleanup_mux to clean the gsm_mux[]
entry. And reactivate it later. If gsmld_open calls in, then the
race may make the orignal gsm->num changed, or gsm_activate_mux
just fails. In both scenarios, we may hit panic.

Change-Id: I08889f038fa67f9e74ccafbcae1067180c03a8d6
Tracked-On: https://jira01.devtools.intel.com/browse/AKP-1778
Signed-off-by: xinhui.pan <xinhuiX.pan@intel.com>
Reviewed-by: Zhang Yanmin <yanmin_zhang@linux.intel.com>
Reviewed-on: https://android.intel.com:443/372495