FOR_UPSTREAM [VPG]: drm/i915 fix per engine reset stats
With a per engine reset (in i915_handle_hung_ring) the
calls to i915_set_reset_status for each non-completed
batch buffer request are all made with guilty=false.
This means that hang_stats->batch_active remains zero,
implying that no batch buffer was involved in the hang.
This causes gem_reset_stats subtest reset-stats-render
(from IGT) to fail.
Modify the calls to i915_set_reset_status so that the
first request in the list that has not completed is
marked as "guilty". This is most likely the problem
batch, and in any case we are just counting stats and
we know that one batch must be responsible for the hang.

Change-Id: Iec5e1a0651e4295a4ec92e0208339539f7e5f207
For: VIZ-5758
Signed-off-by: Tim Gore <tim.gore@intel.com>