Revert "FOR_UPSTREAM [VPG]: drm/i915/gen8: Apply Per-context workarounds using W/A batch buffers"
This reverts the following commit to replace them with upstream patches.
This patch was also sent upstream for review and recently got merged.
Upstream patches are much simpler and contained and we identified that some
of the WA need not be applied so it is better to use upstream patches.

commit 695f598c34ba73f2f517ce43825cfe110cda1eea
Author: Arun Siluvery <arun.siluvery@linux.intel.com>
Date:   Wed Nov 5 19:00:26 2014 +0000

    FOR_UPSTREAM [VPG]: drm/i915/gen8: Apply Per-context workarounds using W/A batch buffers

For: GMINL-3300
For: GMINL-3301
Change-Id: I5f09685608a56e53f6111b390cd340a488f12af3
Signed-off-by: Arun Siluvery <arun.siluvery@linux.intel.com>