typec: resolve the charger disconnect issue with pd charger
when pd charger is connected to the DUT, without connecting it to
the power socket, DUT recognizes it as a peripheral device (as UFP)
and enters to DFP mode supplying 5V.  FUSB does disconnect due to
CC voltage range mismatch.

This patch handles the race conditions to determine proper
connect / disconnect and notify the events accordingly.

If the detection driver receives DFP event, restart drp when valid
CC cant be determined.

Change-Id: Ib757039b0afb584839d017852d24d0724f6537c5
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-13301
Signed-off-by: R, Kannappan <r.kannappan@intel.com>