REVERTME [VPG]: drm/i915: update arbiter in non-atomic case
While in full screen video, i.e only sprite enable case, if sprite
plane is forcefully disabled, as part of sprite disable, primary
plane is enabled. Hence update dsparb also so that some buffer will
be allocated for primary.

This patch is REVERTME because it is based on the VPG internal
set_display atomic call. On migrating to nuclear page flip this will be
reverted.

Tracked-On: http://jira01.devtools.intel.com/browse/IMINAN-41371
Change-Id: I07cd8e24a6a094bea7b5602c86c5ae74b2a48381
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>
Signed-off-by: Shen, Lei <lei.shen@intel.com>
## Don't forget to add a JIRA URL to the Tracked-On metadata!
## https://wiki.ith.intel.com/display/CACTUS/Patch+requirements+check
## url should be in the form:
## Tracked-On: https://jira01.devtools.intel.com/browse/<PROJ>-<NUMB>
## Several Tracked-On lines are allowed
## you can fully prefill your Tracked-On commits using environment variable:
## export CURRENT_TRACKED_ON=https://jira01.devtools.intel.com/browse/<PROJ>-<NUMB>
##