Revert "FOR_UPSTREAM [VPG]: drm/i915: 2nd fix for reset stats"
This reverts commit 181197ebd5d7701379a0b40f12c52a48d2c6a6d1.

The original commit was to aid the OGL robustness interface
in detecting per engine reset when that reset had not
actually affected the context that it was considering.
After some discussion it was decided to revert this commit
and add code in the usermode OGL driver to look at
batch_pending and batch_active counts across all relevant
contexts to determine the impact of per engine resets.
This is almost certainly a better solution.

However, some igt gem_reset_stats subtests were fixed
by this patch, so now they are broken again when TDR is
enabled. There are several subtests that explicitly check
the reset count returned by i915_get_reset_stats_ioctl and
these now fail when TDR is enabled because TDR does not do
a full GPU reset and so does not increment the reset count.
When TDR goes upstream we will need to make the
gem_reset_stats test TDR aware but for now we probably
need to just ignore these subtests.

Change-Id: I05456cfac607bfed4dfd744480d49a1e42fdbf50
Tracked-On: https://jira01.devtools.intel.com/browse/VIZ-5904
Signed-off-by: Tim Gore <tim.gore@intel.com>

## Don't forget to add a JIRA URL to the Tracked-On metadata!
## https://wiki.ith.intel.com/display/CACTUS/Patch+requirements+check
## url should be in the form:
## Tracked-On: https://jira01.devtools.intel.com/browse/<PROJ>-<NUMB>
## Several Tracked-On lines are allowed
## you can fully prefill your Tracked-On commits using environment variable:
## export CURRENT_TRACKED_ON=https://jira01.devtools.intel.com/browse/<PROJ>-<NUMB>
##