FROM_UPSTREAM [VPG]: drm/i915/gtt: Introduce i915_page_dir_dma_addr
The legacy mode mm switch and the execlist context assignment
needs dma address for the page directories.

Introduce a function that encapsulates the scratch_pd dma
fallback if no pd is found.

v2: Rebase, s/ring/req

Note for future forklift, this patch was added in upstream after
the ring/req changes (v2). As these changes are not yet in 3.14,
this was modified to fit.
If drm_i915_gem_request is ever added to this kernel, replace this
patch with the upstreamed v2.

Change-Id: I45ce6b4ae3cb712d7423e378a232c910777bab32
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-11653
Upstream-repo: git://anongit.freedesktop.org/drm-intel
Upstream-commit: d852c7bf902aa36f37b6fd0bb1147c922e7bd0fb
Upstream-as-of-tag: drm-intel-next-2015-07-03
Signed-off-by: Mika Kuoppala <mika.kuoppala@intel.com>
Reviewed-by: Michel Thierry <michel.thierry@intel.com> (v1)
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Michel Thierry <michel.thierry@intel.com>
Reviewed-on: https://android.intel.com:443/445185