Update battery charging safety params from OEM0 table
To charge the battery safely, the max and min temperature should be
based on the profile which is supported by the platform. Currently
the battery min and max temperature values are hardcoded in the
charger driver. This will cause an issue when the charging profile
temp value varies from the hardcoded values. Hence the min and max
temp values are updated based on the OEMO table data.

Change-Id: I2bb5807ece5c6a46260b74e72db113d596c32ae7
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-13901
Signed-off-by: Albin B <albin.bala.krishnan@intel.com>