FOR_UPSTREAM [VPG]: drm/i915: Command mode Support with IMINAN- 39625
To optimize the blank out time during boot, the modeset was by-passed
in driver if already done in the BIOS. In case of command-mode,
the BIOS bu default configuring for video mode, hence modeset
is required in the driver to support command mode.
The check has been added to by-pass modeset only in video mode.

Change-Id: I2a5da9b7cd30dcaf2fa555be48dd16d84fae2377
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7482
Signed-off-by: Megha Aggarwal <megha.aggarwal@intel.com>
Signed-off-by: Ramalingam C <ramalingam.c@intel.com>