Improve cpu load calculation on Intel Platforms
The cpu load computed by the interactive governor is not accurate
because it is adjusted with a frequency (pcpu->policy->cur) which
is different from the real frequency (because of throttling but
also because we have 2 cpus sharing the same frequency).
One can improve that simply by using the APERF/MPERF and TSC counters.

Whenever this patch fails during a kernel rebase on either kernel.org
stable updates or Google common.git updates, contact
philippe.longepe@intel.com to have this patch rebased

Change-Id: Iadc7e69eaf179f4cd610e1d417b2bc33f4230512
Tracked-On: https://jira01.devtools.intel.com/browse/AKP-2252
Signed-off-by: Philippe Longepe <philippe.longepe@intel.com>