Process all the extcon event for vbus connect/disconnect
When plug/unplug iterations, TI device resolves as SNK and FFD as SRC,
however FFD never provides VBUS and communication fails. TI device is
a DRP so it randomly resolved as either DFP or UFP, this happens
during a transition from DFP to UFP. So process all the extcon events
for vbus connect/disconnect.

Change-Id: I42943aa75e0e58b7afc96602098247293fb14685
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-18896
Signed-off-by: Albin B <albin.bala.krishnan@intel.com>