FOR_UPSTREAM [VPG]: drm/i915: Use fail safe mode when edid is corrupt
This patch forces panel to use fail safe mode when it is detected
that edid is corrupt.

For now this is performed only when automated test request is
received. This is because DRM's edid parser recovers from
basic corruption in edid resulting in normal functioning.

This can be modified to be used anytime by using edid_corrupt
flag instead of checking for Atomated test request flag.

v2: use 18bpp when using failsafe mode

Change-Id: I843696f005e519a51394c0834bf0f01e50e53b19
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-9749
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Reviewed-on: https://android.intel.com:443/439698