FOR_UPSTREAM [VPG]: drm/i915: Make pages of GFX allocations movable
On a long run of more than 2-3 days, physical memory tends to get fragmented
severely. This external fragmentation slows down the system so much that
starting a new process with some UI takes inordinate time to grab the
required buffers. In such a scenario, low memory killer or Shrinker is also
unable to help as lack of memory is not the actual problem, since it has been
observed that there are enough free pages of 0 order. It seems there are
requests to allocate physically-contiguous pages which are no longer available
and so triggers compaction but almost in vain.

To address the issue of external fragementation, kernel does a compaction
(which involves migration of pages) but it's efficacy depends upon how many
pages are marked as MOVABLE, as only those pages can be migrated.

Currently the backing pages for GFX buffers are allocated from shmem
with GFP_RECLAIMABLE flag, in units of 4KB pages.
With limited Swap space (ZRAM), it may not be possible always to reclaim/swap-out
pages of all the inactive objects, to make way for free space allowing formation
of higher order groups of physically-contiguous pages on compaction.

The GFX pages cannot be simply marked as MOVABLE, as GFX Driver stores a
reference to the pages (struct page*), which will become invalid if the
pages are migrated.
For that a 'migratepage' callback has been implemented, which is invoked
by kernel just prior to migration, wherein Driver can release the reference
to pages. This allows Driver to mark the GFX pages as MOVABLE and hence mitigate
the fragmentation problem.

v2: Skip page migration for purgeable objects, to avoid the deadlock

v3: Skip migration of objects pinned for Display

v4: Handle the race due to two threads locking page_lock and i915 struct mutex
    in inverse order. Also, there may be a race in which object is freed up
    wherein the pages are still being migrated. Handled these conditions.

v5: Cover the drop_pages with rpm get/put calls.

v6: Lock struct_mutex right at the beginning of i915_migratepage fn.

v7: Clear out the private field of a new page to get rid of the stale value.

v8: Skip migration, when device is not runtime active.

v9: Skip migration and return -EBUSY, when refcount > 2 for a page, which
    is already unreferenced by the Driver.

Change-Id: I37cf470b9555f6445bd9e54814cf0f58181f352c
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-13363
Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>
Signed-off-by: Akash Goel <akash.goel@intel.com>
Reviewed-on: https://android.intel.com:443/439421