REVERTME i2c: designware: debug patch for pmic kernel panic issue
This patch adds debug log for the i2c controller timed out issue.
Christian Processe suggests to put this patch in mainline:
"The point is that the issue is reproduced by SQE validation team mainly
& boards are used for official validation in priority then the boards
have few availability for the debug except if we  can integrate the
patch in mainline."

This patch will be reverted once we capture the issue.

Change-Id: I1f62fc8b5d67568e0ce0af7187e23a11ead12ca8
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-14705
Signed-off-by: Xiang Wang <xiang.a.wang@intel.com>
Reviewed-on: https://android.intel.com:443/457298