iio: magn: ak09911: Optimization of the trigger handler
the data readout frequency was low inspite of using the hrtimer,
also the % jitter observed within the apps was high.
It was observed that we were spending considerable time within
the hrtimer trigger handler in setting the operating mode each time
and waiting for the data ready to happen. Optimized the function by
removing the call to wait_conversion_complete_polled and adding poll
data ready at a faster rate.

Change-Id: I40fe2c3f300d096e2a5d2beedaac788bfcc7e65f
Tracked-On:https://jira01.devtools.intel.com/browse/OAM-28475
Signed-off-by: Arun Radhakrishnan <arun.radhakrishnan@intel.com>
Reviewed-on: https://android.intel.com:443/492688