EM/extcon-dc-pwrsrc: Register extcon notifier on all cases
Issue:
USB host/peripheral mode notification can come in 2 ways.
1. On platforms where USB ID and VBUS are controlled by OTG,
the notification is received from otg driver.
2. On platforms where ID and VBUS are controlled by GPIO,
the notification could be routed through extcon.
In the 2 approach, the mode detection could be done in 3gpio
extcon driver or in this pwrsrc driver itself. Current implementation
would fail if the detection is done in 3gpio extcon driver.

Solution:
Register for the extcon notifications no matter which driver does
the USB host/peripheral mode detection. If the GPIO resources for ID
and VBUS are defined for PMIC device in BIOS, this power source driver
would send the notification and is processed here itself. If the GPIO
resources for ID and VBUS are defined for OTG2 device in BIOS, 3gpio
extcon driver would send the notification and processing is done
here in pwrsrc driver.

This driver has to listen to this notification as to identify whether
the VBUS event was received for ID or for charger connect.

Change-Id: Ia9f6de7e95e91880cfa22f6a4070046bda131838
Signed-off-by: ysiyer <yegnesh.s.iyer@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-30033
Reviewed-on: https://android.intel.com:443/500738