EM/dc_ti_pwrsrc: WA fix to enable the VBUS upon wakeup from S3
Issue:
When the USB peripheral device is connected to the platform
during S3, platform got the IRQ and pwrsrc driver got USB-Host
extcon notification. Soon after, pwrsrc driver is trying to
enable the VBUS through bq charger chip register. But the i2c
bus in which the bq chip is placed has not come out of suspend
at this time. This results in i2c read-write error.

Solution:
Added a delay of 10ms to start the worker thread which enables
the VBUS. By this time i2c bus has completed resume function
and so bq chip access and the VBUS enable is happening
successfully.

Change-Id: I0f97997964dbed82196a86149c17f1f146065714
Signed-off-by: Ameermon V A <ameermonx.va@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-30850
Reviewed-on: https://android.intel.com:443/505552